requirements.txt
django-crispy-forms==1.9.1
psycopg2-binary==2.8.5
python-dateutil==2.8.1
Django==3.0.6
django-nose==1.4.6
coverage==5.3
django-baker==0.22
django-utils-six==2.0
requests==2.25.0
sentry-sdk==0.16.1
django-braces==1.14.0
pulsar-client==2.6.0
django-prometheus==2.1.0
